requirements.txt
python-dotenv==1.0.0
tweepy==4.14.0
web3==6.11.1
requests==2.31.0
python-logging==0.4.9.6
PyYAML==6.0.1